requirements.txt
streamlit>=1.35.0
openai>=1.30.0
langgraph>=0.2.0
streamlit_option_menu 
streamlit_searchbox
yfinance
python-dotenv